moved list into new 'TodoList'component, created new 'AddTodoForm' component, used these elements in 'App' component
